+++ src/Components/Auth/Login.js
@@ -1,5 +1,6 @@
 export default function Login() {
   return (
+    // The proper user authentication functionality will be added at a later date
     <div>
       <h2>Sign In</h2>
       <form>